[TBC] Q1 Part B working
Mean = 0.482160
Estimate the probability P (Zi,1 = 1) that the first column in the dataset takes value 1 i.e. that a shopping basket contains an item 1. Briefly explain/discuss your calculation. Hint: Observe that the first column in the dataset only takes values 0 or 1 and recall that for an indicator RV X we have Prob(X = 1) = E[X]. [5 marks]